Move aggregation functionality into namespace
